// src/generated/t_tech/invest/grpc/operations.ts
export enum OperationState {
  /** OPERATION_STATE_UNSPECIFIED - Статус операции не определен. */
  OPERATION_STATE_UNSPECIFIED = 0,
  /** OPERATION_STATE_EXECUTED - Исполнена частично или полностью. */
  OPERATION_STATE_EXECUTED = 1,
  /** OPERATION_STATE_CANCELED - Отменена. */
  OPERATION_STATE_CANCELED = 2,
  /** OPERATION_STATE_PROGRESS - Исполняется. */
  OPERATION_STATE_PROGRESS = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface GetOperationsByCursorRequest {
  /** Идентификатор счета клиента, обязательный параметр. Остальные параметры опциональны. */
  accountId: string;
  /** Идентификатор инструмента. Принимает значение `figi`, `instrument_uid` или `ticker + '_' + class_code`. */
  instrumentId?:
    | string
    | undefined;
  /** Начало периода по UTC. */
  from?:
    | Date
    | undefined;
  /** Окончание периода по UTC. */
  to?:
    | Date
    | undefined;
  /** Идентификатор элемента, с которого начать формировать ответ. */
  cursor?:
    | string
    | undefined;
  /** Лимит количества операций. По умолчанию — `100`, максимальное значение — `1000`. */
  limit?:
    | number
    | undefined;
  /** Тип операции. Принимает значение из списка `OperationType`. */
  operationTypes: OperationType[];
  /** Статус запрашиваемых операций. Возможные значения указаны в `OperationState`. */
  state?:
    | OperationState
    | undefined;
  /** Флаг возврата комиссии. По умолчанию — `false`. */
  withoutCommissions?:
    | boolean
    | undefined;
  /** Флаг получения ответа без массива сделок. */
  withoutTrades?:
    | boolean
    | undefined;
  /** Флаг показа overnight операций. */
  withoutOvernights?: boolean | undefined;
}

/** Список операций по счету с пагинацией. */
export interface GetOperationsByCursorResponse {
  /** Признак, есть ли следующий элемент. */
  hasNext: boolean;
  /** Следующий курсор. */
  nextCursor: string;
  /** Список операций. */
  items: OperationItem[];
}

/** Данные об операции. */
export interface OperationItem {
  /** Курсор. */
  cursor: string;
  /** Номер счета клиента. */
  brokerAccountId: string;
  /** Идентификатор операции, может меняться с течением времени. */
  id: string;
  /** Идентификатор родительской операции. Может измениться, если изменился ID родительской операции. */
  parentOperationId: string;
  /** Название операции. */
  name: string;
  /** Дата поручения. */
  date:
    | Date
    | undefined;
  /** Тип операции. */
  type: OperationType;
  /** Описание операции. */
  description: string;
  /** Статус поручения. */
  state: OperationState;
  /** Уникальный идентификатор инструмента. */
  instrumentUid: string;
  /** FIGI. */
  figi: string;
  /** Тип инструмента. */
  instrumentType: string;
  /** Тип инструмента. */
  instrumentKind: InstrumentType;
  /** Уникальный идентификатор позиции. */
  positionUid: string;
  /** Тикер инструмента. */
  ticker: string;
  /** Класс-код (секция торгов). */
  classCode: string;
  /** Сумма операции. */
  payment:
    | MoneyValue
    | undefined;
  /** Цена операции за 1 инструмент. */
  price:
    | MoneyValue
    | undefined;
  /** Комиссия. */
  commission:
    | MoneyValue
    | undefined;
  /** Доходность. */
  yield:
    | MoneyValue
    | undefined;
  /** Относительная доходность. */
  yieldRelative:
    | Quotation
    | undefined;
  /** Накопленный купонный доход. */
  accruedInt:
    | MoneyValue
    | undefined;
  /** Количество единиц инструмента. */
  quantity: number;
  /** Неисполненный остаток по сделке. */
  quantityRest: number;
  /** Исполненный остаток. */
  quantityDone: number;
  /** Дата и время снятия заявки. */
  cancelDateTime:
    | Date
    | undefined;
  /** Причина отмены операции. */
  cancelReason: string;
  /** Массив сделок. */
  tradesInfo:
    | OperationItemTrades
    | undefined;
  /** Идентификатор актива. */
  assetUid: string;
  /** Массив дочерних операций. */
  childOperations: ChildOperationItem[];
}

/** Массив с информацией о сделках. */
export interface OperationItemTrades {
  trades: OperationItemTrade[];
}
// ...
export interface ChildOperationItem {
  /** Уникальный идентификатор инструмента. */
  instrumentUid: string;
  /** Сумма операции. */
  payment: MoneyValue | undefined;
}
// ...
export const GetOperationsByCursorResponse: MessageFns<GetOperationsByCursorResponse> = {
  encode(message: GetOperationsByCursorResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.hasNext !== false) {
      writer.uint32(8).bool(message.hasNext);
    }
    if (message.nextCursor !== "") {
      writer.uint32(18).string(message.nextCursor);
    }
    for (const v of message.items) {
      OperationItem.encode(v!, writer.uint32(50).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetOperationsByCursorResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetOperationsByCursorResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.hasNext = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.nextCursor = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.items.push(OperationItem.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetOperationsByCursorResponse {
    return {
      hasNext: isSet(object.hasNext)
        ? globalThis.Boolean(object.hasNext)
        : isSet(object.has_next)
        ? globalThis.Boolean(object.has_next)
        : false,
      nextCursor: isSet(object.nextCursor)
        ? globalThis.String(object.nextCursor)
        : isSet(object.next_cursor)
        ? globalThis.String(object.next_cursor)
        : "",
      items: globalThis.Array.isArray(object?.items) ? object.items.map((e: any) => OperationItem.fromJSON(e)) : [],
    };
  },

  toJSON(message: GetOperationsByCursorResponse): unknown {
    const obj: any = {};
    if (message.hasNext !== false) {
      obj.hasNext = message.hasNext;
    }
    if (message.nextCursor !== "") {
      obj.nextCursor = message.nextCursor;
    }
    if (message.items?.length) {
      obj.items = message.items.map((e) => OperationItem.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<GetOperationsByCursorResponse>): GetOperationsByCursorResponse {
    return GetOperationsByCursorResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetOperationsByCursorResponse>): GetOperationsByCursorResponse {
    const message = createBaseGetOperationsByCursorResponse();
    message.hasNext = object.hasNext ?? false;
    message.nextCursor = object.nextCursor ?? "";
    message.items = object.items?.map((e) => OperationItem.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const OperationItemTrades: MessageFns<OperationItemTrades> = {
  encode(message: OperationItemTrades, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.trades) {
      OperationItemTrade.encode(v!, writer.uint32(50).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OperationItemTrades {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOperationItemTrades();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.trades.push(OperationItemTrade.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OperationItemTrades {
    return {
      trades: globalThis.Array.isArray(object?.trades)
        ? object.trades.map((e: any) => OperationItemTrade.fromJSON(e))
        : [],
    };
  },

  toJSON(message: OperationItemTrades): unknown {
    const obj: any = {};
    if (message.trades?.length) {
      obj.trades = message.trades.map((e) => OperationItemTrade.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<OperationItemTrades>): OperationItemTrades {
    return OperationItemTrades.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<OperationItemTrades>): OperationItemTrades {
    const message = createBaseOperationItemTrades();
    message.trades = object.trades?.map((e) => OperationItemTrade.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const ChildOperationItem: MessageFns<ChildOperationItem> = {
  encode(message: ChildOperationItem, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.instrumentUid !== "") {
      writer.uint32(10).string(message.instrumentUid);
    }
    if (message.payment !== undefined) {
      MoneyValue.encode(message.payment, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChildOperationItem {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChildOperationItem();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.instrumentUid = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.payment = MoneyValue.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ChildOperationItem {
    return {
      instrumentUid: isSet(object.instrumentUid)
        ? globalThis.String(object.instrumentUid)
        : isSet(object.instrument_uid)
        ? globalThis.String(object.instrument_uid)
        : "",
      payment: isSet(object.payment) ? MoneyValue.fromJSON(object.payment) : undefined,
    };
  },

  toJSON(message: ChildOperationItem): unknown {
    const obj: any = {};
    if (message.instrumentUid !== "") {
      obj.instrumentUid = message.instrumentUid;
    }
    if (message.payment !== undefined) {
      obj.payment = MoneyValue.toJSON(message.payment);
    }
    return obj;
  },

  create(base?: DeepPartial<ChildOperationItem>): ChildOperationItem {
    return ChildOperationItem.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ChildOperationItem>): ChildOperationItem {
    const message = createBaseChildOperationItem();
    message.instrumentUid = object.instrumentUid ?? "";
    message.payment = (object.payment !== undefined && object.payment !== null)
      ? MoneyValue.fromPartial(object.payment)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}